requirements.txt
Babel==2.4.0
Django==1.10.7
django-phonenumber-field==1.3.0
phonenumberslite==8.5.2
pytz==2017.2
